requirements.txt
fastapi
uvicorn
python-multipart
requests
google-generativeai
pvfalcon
ffmpeg